requirements.txt
transformers
datasets
evaluate
torch
numpy
scipy
edit_distance
praat-textgrids
speechbrain